Добавлен билд в библиотеку
