+++ lib/designTokens/tokens/index.ts
@@ -1,0 +1,3 @@
+export { crmTokens } from './crmTokens.ts';
+export { joyDeckTokens } from './joyDeckTokens.ts';
+export { legalKZTokens } from './legalKZTokens.ts';
