+++ lib/main.ts
@@ -1,0 +1,4 @@
+export * from './designTokens/tokens';
+
+export { parseDesignTokens } from './designTokens/utils/parseDesignTokens';
+
