+++ package.json
@@ -1,38 +1,48 @@
 {
-  "name": "ant-design-system",
-  "private": true,
-  "version": "0.0.0",
-  "type": "module",
-  "scripts": {
-    "dev": "vite",
-    "build": "tsc -b && vite build",
-    "preview": "vite preview"
-  },
-  "dependencies": {
-    "@ant-design/icons": "^5.4.0",
-    "@tanstack/react-table": "^8.7.6",
-    "@tanstack/table-core": "^8.20.1",
-    "antd": "^5.20.2",
-    "classnames": "^2.5.1",
-    "dayjs": "^1.11.13",
-    "react": "^18.3.1",
-    "react-dom": "^18.3.1",
-    "react-hook-form": "^7.53.0",
-    "react-icons": "^5.3.0",
-    "sass": "^1.77.8",
-    "vite-plugin-svgr": "^4.2.0"
-  },
-  "devDependencies": {
-    "@eslint/js": "^9.9.0",
-    "@types/react": "^18.3.3",
-    "@types/react-dom": "^18.3.0",
-    "@vitejs/plugin-react": "^4.3.1",
-    "eslint": "^9.9.0",
-    "eslint-plugin-react-hooks": "^5.1.0-rc.0",
-    "eslint-plugin-react-refresh": "^0.4.9",
-    "globals": "^15.9.0",
-    "typescript": "^5.5.3",
-    "typescript-eslint": "^8.0.1",
-    "vite": "^5.4.1"
-  }
+    "name": "ant-design-system",
+    "private": true,
+    "version": "0.0.0",
+    "type": "module",
+    "main": "dist/main.js",
+    "types": "dist/main.d.ts",
+    "scripts": {
+        "dev": "vite",
+        "build": "tsc --p ./tsconfig-build.json && vite build",
+        "preview": "vite preview"
+    },
+    "dependencies": {
+        "@tanstack/react-table": "^8.7.6",
+        "@tanstack/table-core": "^8.20.1",
+        "classnames": "^2.5.1",
+        "dayjs": "^1.11.13",
+        "react-hook-form": "^7.53.0",
+        "react-icons": "^5.3.0",
+        "sass": "^1.77.8"
+    },
+    "peerDependencies": {
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "@ant-design/icons": "^5.4.0",
+        "antd": "^5.20.2"
+    },
+    "devDependencies": {
+        "@eslint/js": "^9.9.0",
+        "@types/node": "^22.9.0",
+        "@types/react": "^18.3.3",
+        "@types/react-dom": "^18.3.0",
+        "@vitejs/plugin-react": "^4.3.1",
+        "react": "^18.3.1",
+        "react-dom": "^18.3.1",
+        "@ant-design/icons": "^5.4.0",
+        "antd": "^5.20.2",
+        "eslint": "^9.9.0",
+        "eslint-plugin-react-hooks": "^5.1.0-rc.0",
+        "eslint-plugin-react-refresh": "^0.4.9",
+        "globals": "^15.9.0",
+        "typescript": "^5.5.3",
+        "typescript-eslint": "^8.0.1",
+        "vite": "^5.4.1",
+        "vite-plugin-dts": "^4.3.0",
+        "vite-plugin-svgr": "^4.2.0"
+    }
 }
