+++ tsconfig-build.json
@@ -1,0 +1,6 @@
+{
+    "extends": "./tsconfig.app.json",
+    "include": [
+        "lib"
+    ]
+}
