+++ vite.config.ts
@@ -1,8 +1,25 @@
 import { defineConfig } from 'vite';
 import react from '@vitejs/plugin-react';
 import svgr from 'vite-plugin-svgr';
+import dts from 'vite-plugin-dts';
+import { resolve } from 'path';
 
 // https://vitejs.dev/config/
 export default defineConfig({
-  plugins: [react(), svgr()],
+  plugins: [react(), svgr(), dts({
+    include: ['lib'],
+
+    tsconfigPath: './tsconfig.app.json',
+  })],
+  build: {
+    copyPublicDir: false,
+    lib: {
+      fileName: 'main',
+      entry: resolve(__dirname, 'lib/main.ts'),
+      formats: ['es'],
+    },
+    rollupOptions: {
+      external: ['react', 'react/jsx-runtime'],
+    },
+  },
 });
